# Daesc_Back/swagger_documentos.json
{
  "swagger": "2.0",
  "info": {
    "title": "Minhas rotas",
    "descrição": "Documentação do projeto daesc",
    "version": "1.0.0",
    "description": ""
  },
  "host": "localhost:8083",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/admin/situacaoSimulacao/deletar": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/admin/situacaoReclamacao/deletar": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/admin/criar/novo": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "example": "any"
                },
                "sobrenome": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "senha": {
                  "example": "any"
                },
                "senha2": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
...
    "/user/servico/novo": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "boleto": {
                  "example": "any"
                },
                "entrada": {
                  "example": "any"
                },
                "parcela": {
                  "example": "any"
                },
                "data": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/user/situacao/nova": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "rua": {
                  "example": "any"
                },
                "bairro": {
                  "example": "any"
                },
                "descricao": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
...
  }
}